+++ SPPMath/SPPMath.h
@@ -431,19 +431,19 @@ namespace SPP
         // pitch
         double sinr_cosp = 2 * (w * x + y * z);
         double cosr_cosp = 1 - 2 * (x * x + y * y);
-        angles[0] = std::atan2(sinr_cosp, cosr_cosp);
+        angles[0] = (float)std::atan2(sinr_cosp, cosr_cosp);
 
         // yaw
         double sinp = 2 * (w * y - z * x);
         if (std::abs(sinp) >= 1)
-            angles[1] = std::copysign(M_PI / 2, sinp); // use 90 degrees if out of range
+            angles[1] = (float)std::copysign(M_PI / 2, sinp); // use 90 degrees if out of range
         else
-            angles[1] = std::asin(sinp);
+            angles[1] = (float)std::asin(sinp);
 
         // roll
         double siny_cosp = 2 * (w * z + x * y);
         double cosy_cosp = 1 - 2 * (y * y + z * z);
-        angles[2] = std::atan2(siny_cosp, cosy_cosp);
+        angles[2] = (float)std::atan2(siny_cosp, cosy_cosp);
 
         return angles;
     }
